+++ packages/actionsflow-core/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/actionsflow/actionsflow/compare/actionsflow-core@1.4.0...actionsflow-core@1.4.1) (2020-10-31)
+
+
+### Bug Fixes
+
+* 🐛 template epression includes multiple on ([ac3f7b4](https://github.com/actionsflow/actionsflow/commit/ac3f7b49bbb505f08c3d1dfb0782947b0f78bd76))
+
+
+
+
+
+# [1.4.0](https://github.com/actionsflow/actionsflow/compare/actionsflow-core@1.3.2...actionsflow-core@1.4.0) (2020-10-31)
+
+
+### Features
+
+* 🎸 support config filterScript, sortScript ([b2999ea](https://github.com/actionsflow/actionsflow/commit/b2999eabff84182fb7b1d984dcdf304a9a749b1e))
+
+
+
+
+
 ## [1.3.2](https://github.com/actionsflow/actionsflow/compare/actionsflow-core@1.3.1...actionsflow-core@1.3.2) (2020-10-31)