+++ packages/actionsflow-trigger-weather/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.1](https://github.com/actionsflow/actionsflow/compare/@actionsflow/trigger-weather@1.4.0...@actionsflow/trigger-weather@1.4.1) (2020-10-31)
+
+**Note:** Version bump only for package @actionsflow/trigger-weather
+
+
+
+
+
+# [1.4.0](https://github.com/actionsflow/actionsflow/compare/@actionsflow/trigger-weather@1.3.2...@actionsflow/trigger-weather@1.4.0) (2020-10-31)
+
+**Note:** Version bump only for package @actionsflow/trigger-weather
+
+
+
+
+
 ## [1.3.2](https://github.com/actionsflow/actionsflow/compare/@actionsflow/trigger-weather@1.3.1...@actionsflow/trigger-weather@1.3.2) (2020-10-31)
 
 **Note:** Version bump only for package @actionsflow/trigger-weather